Added survey of the Nexmo API.
This is a little bit special because the Nexmo API actually uses a URI segment
to define the data transfer formats (at least for some endpoints). This can be
routed in different ways depending on whether you want to use the same
responder for both formats (decided e.g. by middleware), or have separate
responders defined by two separate routes. For the benchmark to be as complete
as possible, both flavors have been included.